+++ website/docusaurus.config.ts
@@ -8,20 +8,20 @@ import rehypeKatex from 'rehype-katex';
 
 
 const config: Config = {
-  title: 'OPA on AWS',
+  title: 'Harmonix on AWS',
   // tagline: 'Orchestrate Platforms and Applications',
   tagline: 'Fast, secure, and at-scale. A developer portal to meet your Enterprise needs.',
   favicon: 'img/favicon.ico',
 
   // Set the production url of your site here
-  url: 'https://opaonaws.io',
+  url: 'https://harmonixonaws.io',
   // Set the /<baseUrl>/ pathname under which your site is served
   // For GitHub pages deployment, it is often '/<projectName>/'
   baseUrl: '/',
 
   // GitHub pages deployment config.
   // If you aren't using GitHub pages, you don't need these.
-  projectName: 'app-development-for-backstage-io-on-aws', // Usually your repo name.
+  projectName: 'harmonix', // Usually your repo name.
   organizationName: 'awslabs', // Usually your GitHub org/user name.
   onBrokenLinks: 'throw',
   onBrokenMarkdownLinks: 'warn',
@@ -62,15 +62,15 @@ const config: Config = {
 
   themeConfig: {
       // Replace with your project's social card
-    image: 'img/opa-logo.png',
+    image: 'img/harmonix-blue.png',
     navbar: {
-      // title: 'OPA on AWS', 
+      // title: 'Harmonix on AWS', 
       hideOnScroll: false,
       logo: {
-        alt: 'OPA on AWS Logo',
-        src: 'img/opa_white.svg',
-        srcDark: 'img/opa_dark.svg',
-        className: 'opa-nav-logo',
+        alt: 'Harmonix on AWS Logo',
+        src: 'img/harmonix-white.svg',
+        srcDark: 'img/harmonix-white.svg',
+        className: 'harmonix-nav-logo',
         // width: 128,
         // height: 128,
       },
@@ -103,7 +103,7 @@ const config: Config = {
           position: 'left'
         },
         {
-          href: 'https://github.com/awslabs/app-development-for-backstage-io-on-aws',
+          href: 'https://github.com/awslabs/harmonix',
           position: 'right',
           className: 'header-github-link',
           "aria-label": 'Github repository',
@@ -112,7 +112,7 @@ const config: Config = {
     },
     announcementBar: {
       content:
-      'OPA on AWS 0.3.4 is now available.  Check it out and give it a star on <a target="_blank" rel="noopener noreferrer" href="https://github.com/awslabs/app-development-for-backstage-io-on-aws">GitHub</a>! ⭐️.  View changes <a target="_blank" rel="noopener noreferrer" href="https://opaonaws.io/docs/CHANGELOG">here</a>.',
+      'Harmonix on AWS 0.3.4 is now available.  Check it out and give it a star on <a target="_blank" rel="noopener noreferrer" href="https://github.com/awslabs/harmonix">GitHub</a>! ⭐️.  View changes <a target="_blank" rel="noopener noreferrer" href="https://harmonixonaws.io/docs/CHANGELOG">here</a>.',
       backgroundColor: '#ffffff',
       textColor: '#000000',
       isCloseable: true,
@@ -151,7 +151,7 @@ const config: Config = {
             // },
             {
               label: 'GitHub',
-              href: 'https://github.com/awslabs/app-development-for-backstage-io-on-aws',
+              href: 'https://github.com/awslabs/harmonix',
             },
           ],
         },
@@ -164,7 +164,7 @@ const config: Config = {
       respectPrefersColorScheme: false,
     },
     zoom: {
-      selector: 'img:not(.opa-nav-logo)',
+      selector: 'img:not(.harmonix-nav-logo)',
       background: {
         light: 'rgb(255, 255, 255)',
         dark: 'rgb(50, 50, 50)'
@@ -191,7 +191,7 @@ const config: Config = {
     ['docusaurus-plugin-remote-content',  
     {
       name: "content",
-      sourceBaseUrl: "https://raw.githubusercontent.com/awslabs/app-development-for-backstage-io-on-aws/main/",
+      sourceBaseUrl: "https://raw.githubusercontent.com/awslabs/harmonix/main/",
       outDir: "docs",
       documents: ["CHANGELOG.md"],
       performCleanup: true,