+++ website/src/pages/index.tsx
@@ -8,24 +8,29 @@ import Layout from '@theme/Layout';
 import HomepageFeatures from '@site/src/components/HomepageFeatures';
 import Grid from '@mui/material/Grid';
 import styles from './index.module.css';
+
 // import 'bootstrap/dist/css/bootstrap.css';
 
 function HomepageHeader() {
   const {siteConfig} = useDocusaurusContext();
   return (
-    <header className={clsx('hero hero--primary', styles.heroBanner)}>
+    <header className={clsx('hero hero--primary', styles.heroBanner)} style={{paddingTop:'0'}}>
     {/* <header className={clsx('hero hero--primary', styles.heroBanner)} style={{backgroundImage: `url('/img/bg_splash.png')`}} > */}
         {/* <div className='bg-splash'>
           <img src={useBaseUrl('img/bg_splash.svg')} />
-        </div> */}
-      <Grid container spacing={3} className="perspective-container">
-        <Grid item xs={6} className="hero hero--primary bgimg logoWrapper">
-          <span className='logoText1'>Orchestrate Platforms <br/>and Applications</span>
+        </div> */}  
+
+      <Grid container spacing={3} className="perspective-container" style={{marginLeft:'0px'}}>
+        <Grid container spacing={0} item xs={12} direction="column" alignItems="center" display="flex" justifyContent="center" style={{padding:'0'}}>
+              <span style={{color:'#ff9900', fontWeight:'bold', fontStyle:'italic', position:'absolute', top:'5px', zIndex:'1000'}}>📢 OPA on AWS is now "Harmonix on AWS"📢</span>
+          </Grid>
+        <Grid item xs={5} className="hero hero--primary bgimg logoWrapper">
+          {/* <span className='logoText1'>Orchestrate Platforms <br/>and Applications</span> */}
           <span className='logoText2'>on AWS</span>
         </Grid>
-        <Grid className='tagWrapper' item xs={5}>
-          <Grid container direction="column" spacing={2} style={{display: 'flex'}}>
-            <Grid item  className="hero__subtitle tagline" style={{display: 'flex', alignItems: 'left', marginTop: 50, paddingRight: 50}}>
+        <Grid className='tagWrapper' item xs={7}>
+          <Grid container direction="column" spacing={2}>
+            <Grid item  className="hero__subtitle tagline" style={{ marginTop: 50, paddingLeft:'50px'}}>
               <p>{siteConfig.tagline}</p>
             </Grid>
             {/* <Grid item style={{alignItems: 'left', textAlign: "center", paddingRight: 50}}>
@@ -37,6 +42,9 @@ function HomepageHeader() {
           <Grid item xs={3}/>
         </Grid>
       </Grid>
+      {/* <Row container spacing={3}>
+    
+      </Row> */}
     </header>
   );
 }